+++ apps/oxlint/scripts/build.js
@@ -34,8 +34,6 @@ const parserFilePaths = [
   */
   'generated/deserialize/ts_range_loc_parent_no_parens.js',
   'generated/visit/keys.js',
-  'generated/visit/types.js',
-  'generated/visit/walk.js',
 ];
 
 for (const parserFilePath of parserFilePaths) {
