+++ apps/oxlint/test/compile-visitor.test.ts
@@ -3,11 +3,11 @@ import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 /*
 // TODO(camc314): we need to generate `.d.ts` file for this module.
 // @ts-expect-error
-import { NODE_TYPE_IDS_MAP } from '../dist/generated/lazy/types.js';
+import { NODE_TYPE_IDS_MAP } from '../src-js/generated/types.js';
 */
 // TODO(camc314): we need to generate `.d.ts` file for this module
 // @ts-expect-error
-import { NODE_TYPE_IDS_MAP } from '../dist/generated/visit/types.js';
+import { NODE_TYPE_IDS_MAP } from '../src-js/generated/types.js';
 import {
   addVisitorToCompiled,
   compiledVisitor,
